+++ pages/space/[slug]/index.tsx
@@ -1,274 +1,16 @@
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import { SpaceContext } from '@lib/context';
-import { useCreateList, useCreateProperty, useFindManyList, useFindManyProperty } from '@lib/hooks';
+import { useFindManyList, useFindManyProperty } from '@lib/hooks';
 import { List, Space, User } from '@prisma/client';
-import { PropertyType } from '@zenstackhq/runtime/models';
 import BreadCrumb from 'components/BreadCrumb';
 import SpaceMembers from 'components/SpaceMembers';
 import TodoList from 'components/TodoList';
 import WithNavBar from 'components/WithNavBar';
 import { GetServerSideProps } from 'next';
 import { useRouter } from 'next/router';
-import { ChangeEvent, FormEvent, useContext, useRef, useState } from 'react';
-import { toast } from 'react-toastify';
 import { getEnhancedPrisma } from 'server/enhanced-db';
-import { LockClosedIcon } from '@heroicons/react/24/outline';
 import PropertyList from 'components/Property/PropertyList';
+import { CreateListDialog } from 'components/List/CreateListDialog';
+import { CreatePropertyDialog } from 'components/Property/CreatePropertyDialog';
 
-function CreateListDialog() {
-    const space = useContext(SpaceContext);
-
-    const [modalOpen, setModalOpen] = useState(false);
-    const [title, setTitle] = useState('');
-    const [_private, setPrivate] = useState(false);
-
-    const create = useCreateList();
-
-    const inputRef = useRef<HTMLInputElement>(null);
-
-    const onSubmit = async (event: FormEvent) => {
-        event.preventDefault();
-
-        try {
-            await create.mutateAsync({
-                data: {
-                    title,
-                    private: _private,
-                    space: { connect: { id: space!.id } },
-                },
-            });
-        } catch (err: any) {
-            toast.error(`Failed to create list: ${err.info?.message || err.message}`);
-            return;
-        }
-
-        toast.success('List created successfully!');
-
-        // reset states
-        setTitle('');
-        setPrivate(false);
-
-        // close modal
-        setModalOpen(false);
-    };
-
-    return (
-        <>
-            <input
-                type="checkbox"
-                id="create-list-modal"
-                className="modal-toggle"
-                checked={modalOpen}
-                onChange={(e: ChangeEvent<HTMLInputElement>) => {
-                    setModalOpen(e.currentTarget.checked);
-                }}
-            />
-            <div className="modal">
-                <div className="modal-box">
-                    <h3 className="font-bold text-xl mb-8">Create a Todo list</h3>
-                    <form onSubmit={(e) => void onSubmit(e)}>
-                        <div className="flex flex-col space-y-4">
-                            <div className="flex items-center">
-                                <label htmlFor="title" className="text-lg inline-block w-20">
-                                    Title
-                                </label>
-                                <input
-                                    id="title"
-                                    type="text"
-                                    required
-                                    placeholder="Title of your list"
-                                    ref={inputRef}
-                                    className="input input-bordered w-full max-w-xs mt-2"
-                                    value={title}
-                                    onChange={(e: FormEvent<HTMLInputElement>) => setTitle(e.currentTarget.value)}
-                                />
-                            </div>
-                            <div className="flex items-center">
-                                <label htmlFor="private" className="text-lg inline-block w-20">
-                                    Private
-                                </label>
-                                <input
-                                    id="private"
-                                    type="checkbox"
-                                    className="checkbox"
-                                    onChange={(e: FormEvent<HTMLInputElement>) => setPrivate(e.currentTarget.checked)}
-                                />
-                            </div>
-                        </div>
-                        <div className="modal-action">
-                            <input className="btn btn-primary" type="submit" value="Create" />
-                            <label htmlFor="create-list-modal" className="btn btn-outline">
-                                Cancel
-                            </label>
-                        </div>
-                    </form>
-                </div>
-            </div>
-        </>
-    );
-}
-
-
-function CreatePropertyDialog() {
-    const space = useContext(SpaceContext);
-
-    const [modalOpen, setModalOpen] = useState(false);
-    const [type, setType] = useState<PropertyType>('APARTMENT');
-    const [address, setAddress] = useState('');
-    const [city, setCity] = useState('');
-    const [postalCode, setPostalCode] = useState('');
-    const [country, setCountry] = useState('');
-    const [_private, setPrivate] = useState(false);
-
-    const create = useCreateProperty();
-
-    const onSubmit = async (event: FormEvent) => {
-        event.preventDefault();
-
-        try {
-            await create.mutateAsync({
-                data: {
-                    type,
-                    address,
-                    city,
-                    postalCode,
-                    country,
-                    private: _private,
-                    space: { connect: { id: space!.id } },
-                },
-            });
-        } catch (err: any) {
-            toast.error(`Failed to create property: ${err.info?.message || err.message}`);
-            return;
-        }
-
-        toast.success('Property created successfully!');
-
-        // reset states
-        setType('APARTMENT');
-        setAddress('');
-        setCity('');
-        setPostalCode('');
-        setCountry('');
-        setPrivate(false);
-
-        // close modal
-        setModalOpen(false);
-    };
-
-    return (
-        <>
-            <input
-                type="checkbox"
-                id="create-property-modal"
-                className="modal-toggle"
-                checked={modalOpen}
-                onChange={(e: ChangeEvent<HTMLInputElement>) => {
-                    setModalOpen(e.currentTarget.checked);
-                }}
-            />
-            <div className="modal">
-                <div className="modal-box">
-                    <h3 className="font-bold text-xl mb-8">Create a Property</h3>
-                    <form onSubmit={(e) => void onSubmit(e)}>
-                        <div className="flex flex-col space-y-4">
-                            <div className="flex items-center">
-                                <label htmlFor="type" className="text-lg inline-block w-20">
-                                    Type
-                                </label>
-                                <select
-                                    id="type"
-                                    required
-                                    className="select select-bordered w-full max-w-xs mt-2"
-                                    value={type}
-                                    onChange={(e: ChangeEvent<HTMLSelectElement>) => setType(e.currentTarget.value as PropertyType)}
-                                >
-                                    <option value="APARTMENT">Apartment</option>
-                                    <option value="HOUSE">House</option>
-                                    <option value="COMMERCIAL">Commercial</option>
-                                </select>
-                            </div>
-                            <div className="flex items-center">
-                                <label htmlFor="address" className="text-lg inline-block w-20">
-                                    Address
-                                </label>
-                                <input
-                                    id="address"
-                                    type="text"
-                                    required
-                                    placeholder="Address of the property"
-                                    className="input input-bordered w-full max-w-xs mt-2"
-                                    value={address}
-                                    onChange={(e: FormEvent<HTMLInputElement>) => setAddress(e.currentTarget.value)}
-                                />
-                            </div>
-                            <div className="flex items-center">
-                                <label htmlFor="city" className="text-lg inline-block w-20">
-                                    City
-                                </label>
-                                <input
-                                    id="city"
-                                    type="text"
-                                    required
-                                    placeholder="City"
-                                    className="input input-bordered w-full max-w-xs mt-2"
-                                    value={city}
-                                    onChange={(e: FormEvent<HTMLInputElement>) => setCity(e.currentTarget.value)}
-                                />
-                            </div>
-                            <div className="flex items-center">
-                                <label htmlFor="postalCode" className="text-lg inline-block w-20">
-                                    Postal Code
-                                </label>
-                                <input
-                                    id="postalCode"
-                                    type="text"
-                                    required
-                                    placeholder="Postal Code"
-                                    className="input input-bordered w-full max-w-xs mt-2"
-                                    value={postalCode}
-                                    onChange={(e: FormEvent<HTMLInputElement>) => setPostalCode(e.currentTarget.value)}
-                                />
-                            </div>
-                            <div className="flex items-center">
-                                <label htmlFor="country" className="text-lg inline-block w-20">
-                                    Country
-                                </label>
-                                <input
-                                    id="country"
-                                    type="text"
-                                    required
-                                    placeholder="Country"
-                                    className="input input-bordered w-full max-w-xs mt-2"
-                                    value={country}
-                                    onChange={(e: FormEvent<HTMLInputElement>) => setCountry(e.currentTarget.value)}
-                                />
-                            </div>
-                            <div className="flex items-center">
-                                <label htmlFor="private" className="text-lg inline-block w-20">
-                                    Private
-                                </label>
-                                <input
-                                    id="private"
-                                    type="checkbox"
-                                    className="checkbox"
-                                    onChange={(e: FormEvent<HTMLInputElement>) => setPrivate(e.currentTarget.checked)}
-                                />
-                            </div>
-                        </div>
-                        <div className="modal-action">
-                            <input className="btn btn-primary" type="submit" value="Create" />
-                            <label htmlFor="create-property-modal" className="btn btn-outline">
-                                Cancel
-                            </label>
-                        </div>
-                    </form>
-                </div>
-            </div>
-        </>
-    );
-}
 
 type Props = {
     space: Space;
